refactor(frontend): split model instance rendering
